inscripcion_backend/requirements/base.txt
Django==4.2.9
psycopg2-binary==2.9.9
graphene-django==3.2.0
django-cors-headers==4.3.1
django-filter==23.5
python-decouple==3.8
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
dj-database-url==2.1.0
